Wrap objects with scope to have names for doc, remove OptimizationStatistics
