+++ lib/AccessorReader.js
@@ -13,12 +13,12 @@ module.exports = AccessorReader;
 /**
  * Reads an accessor incrementally. This is useful to keep overhead low when
  * memory consumption is important.
+ * glTF must be initialized for the pipeline.
  *
  * @param {Object} gltf A javascript object holding a glTF hierarchy.
  * @param {Object} accessor The accessor object from glTF to read.
  * @constructor
  *
- * gltf must be initialized for the pipeline.
  * @see addPipelineExtras
  * @see loadGltfUris
  */
@@ -99,7 +99,7 @@ AccessorReader.prototype.write = function(data, componentType, dataOffset) {
 /**
  * Get if the AccessorReader is at the end of the accessor data.
  *
- * @returns {Boolean} true if there is more data to read, false if not.
+ * @returns {Boolean} True if there is more data to read, false if not.
  */
 AccessorReader.prototype.hasNext = function() {
     return this.index < this.count;
