+++ lib/AmbientOcclusion.js
@@ -24,7 +24,7 @@ var NodeHelpers = require('./NodeHelpers');
 var readAccessor = require('./readAccessor');
 var StaticUniformGrid = require('./StaticUniformGrid');
 
-module.exports = {
+/*module.exports = {
     bakeAmbientOcclusion: bakeAmbientOcclusion,
     // For unit testing:
     generateOptions: generateOptions,
@@ -34,7 +34,7 @@ module.exports = {
     raytraceOverTriangleSamples: raytraceOverTriangleSamples,
     extractInstructionWithFunctionCall: extractInstructionWithFunctionCall,
     injectGlslAfterInstructionContaining: injectGlslAfterInstructionContaining
-};
+};*/
 
 var scratchRay = new Ray();
 var barycentricCoordinateScratch = new Cartesian3();
@@ -46,6 +46,14 @@ var matrix3Scratch = new Matrix3();
 var matrix4Scratch = new Matrix4();
 var trianglePixelMarchScratch = new TrianglePixelMarchOptions();
 var computeAmbientOcclusionAtScratch = new ComputeAmbientOcclusionAtOptions();
+
+module.exports = AmbientOcclusion;
+
+/**
+ * Pipeline stage and utilities for generating ambient occlusion.
+ * @constructor
+ */
+function AmbientOcclusion() {}
 
 /**
  * Bakes ambient occlusion (AO) to all models in a specific scene of the gltf.
@@ -75,7 +83,7 @@ var computeAmbientOcclusionAtScratch = new ComputeAmbientOcclusionAtOptions();
  * @see addPipelineExtras
  * @see loadGltfUris
  */
-function bakeAmbientOcclusion(aoOptions) {
+AmbientOcclusion.bakeAmbientOcclusion = function(aoOptions) {
     var options = generateOptions(aoOptions);
     var gltf = aoOptions.gltfWithExtras;
 
@@ -112,7 +120,16 @@ function bakeAmbientOcclusion(aoOptions) {
             ambientShadowContribution : options.ambientShadowContribution
         });
     }
-}
+};
+
+AmbientOcclusion.generateOptions = generateOptions;
+AmbientOcclusion.generateRaytracerScene = generateRaytracerScene;
+AmbientOcclusion.computeAmbientOcclusionAt = computeAmbientOcclusionAt;
+AmbientOcclusion.raytraceAtTriangleCenters = raytraceAtTriangleCenters;
+AmbientOcclusion.raytraceOverTriangleSamples = raytraceOverTriangleSamples;
+AmbientOcclusion.extractInstructionWithFunctionCall = extractInstructionWithFunctionCall;
+AmbientOcclusion.injectGlslAfterInstructionContaining = injectGlslAfterInstructionContaining;
+
 
 function generateOptions(aoOptions) {
     if (!defined(aoOptions.gltfWithExtras)) {
