+++ lib/GeometryMath.js
@@ -1,16 +1,18 @@
 'use strict';
 var Cesium = require('cesium');
+
 var Cartesian3 = Cesium.Cartesian3;
 var defined = Cesium.defined;
 var DeveloperError = Cesium.DeveloperError;
 
-module.exports = {
-    sumCartesian3sBarycentric: sumCartesian3sBarycentric,
-    flattenTriangle: flattenTriangle,
-    pointLineDistanceParametric: pointLineDistanceParametric
-};
+module.exports = GeometryMath;
 
-// Functions for working with geometry like triangles and lines
+/**
+ * Functions for working with geometry, like triangles and lines.
+ * @constructor
+ * @private
+ */
+function GeometryMath() {}
 
 var sumBarycentricScratch = new Cartesian3();
 /**
@@ -23,7 +25,7 @@ var sumBarycentricScratch = new Cartesian3();
  * @param {Cartesian3} result The object onto which to store the result.
  * @returns {Cartesian3} The modified result parameter.
  */
-function sumCartesian3sBarycentric(barycentric, vector0, vector1, vector2, result) {
+GeometryMath.sumCartesian3sBarycentric = function(barycentric, vector0, vector1, vector2, result) {
     if (!defined(barycentric)) {
         throw new DeveloperError('barycentric is required');
     }
@@ -50,7 +52,7 @@ function sumCartesian3sBarycentric(barycentric, vector0, vector1, vector2, resul
     Cartesian3.multiplyByScalar(vector2, barycentric.z, sumBarycentricScratch);
     Cartesian3.add(result, sumBarycentricScratch, result);
     return result;
-}
+};
 
 var pointOnLineScratch = new Cartesian3();
 /**
@@ -62,7 +64,7 @@ var pointOnLineScratch = new Cartesian3();
  * @param {Cartesian3} lineDirection A direction specifying the line's orientation.
  * @returns {Number} The parametric position on the line closest to the given point.
  */
-function pointLineDistanceParametric(point, linePosition, lineDirection) {
+GeometryMath.pointLineDistanceParametric = function(point, linePosition, lineDirection) {
     if (!defined(point)) {
         throw new DeveloperError('point is required');
     }
@@ -78,7 +80,7 @@ function pointLineDistanceParametric(point, linePosition, lineDirection) {
     var p3 = point;
     var length2 = Cartesian3.magnitudeSquared(lineDirection);
     return ((p3.x - p1.x) * (p2.x - p1.x) + (p3.y - p1.y) * (p2.y - p1.y) + (p3.z - p1.z) * (p2.z - p1.z)) / length2;
-}
+};
 
 var xAxis = new Cartesian3();
 var zAxis = new Cartesian3();
@@ -90,7 +92,7 @@ var yAxis = new Cartesian3();
  * @param {Cartesian2[]} results The objects onto which to store the results.
  * @returns {Cartesian2[]} The triangle's vertex positions in its own plane.
  */
-function flattenTriangle(positions, results) {
+GeometryMath.flattenTriangle = function(positions, results) {
     if (!defined(positions)) {
         throw new DeveloperError('positions is required');
     }
@@ -111,10 +113,10 @@ function flattenTriangle(positions, results) {
     // Compute x/y coordinates by getting distance to normalized x and y axes in 3D
     results[0].x = 0.0;
     results[0].y = 0.0;
-    results[1].x = pointLineDistanceParametric(positions[1], positions[0], xAxis);
-    results[1].y = pointLineDistanceParametric(positions[1], positions[0], yAxis);
-    results[2].x = pointLineDistanceParametric(positions[2], positions[0], xAxis);
-    results[2].y = pointLineDistanceParametric(positions[2], positions[0], yAxis);
+    results[1].x = GeometryMath.pointLineDistanceParametric(positions[1], positions[0], xAxis);
+    results[1].y = GeometryMath.pointLineDistanceParametric(positions[1], positions[0], yAxis);
+    results[2].x = GeometryMath.pointLineDistanceParametric(positions[2], positions[0], xAxis);
+    results[2].y = GeometryMath.pointLineDistanceParametric(positions[2], positions[0], yAxis);
 
     return results;
-}
+};
