+++ lib/Pipeline.js
@@ -3,10 +3,10 @@ var Cesium = require('cesium');
 var Promise = require('bluebird');
 var path = require('path');
 
-var OptimizationStatistics = require('./OptimizationStatistics');
+var AmbientOcclusion = require('./AmbientOcclusion');
+var RemoveUnusedElements = require('./RemoveUnusedElements');
 var addDefaults = require('./addDefaults');
 var addPipelineExtras = require('./addPipelineExtras');
-var bakeAmbientOcclusion = require('./bakeAmbientOcclusion').bakeAmbientOcclusion;
 var cacheOptimization = require('./cacheOptimization');
 var combineMeshes = require('./combineMeshes');
 var combineNodes = require('./combineNodes');
@@ -22,7 +22,6 @@ var mergeDuplicateAccessors = require('./mergeDuplicateAccessors');
 var octEncodeNormals = require('./octEncodeNormals');
 var readGltf = require('./readGltf');
 var removeDuplicatePrimitives = require('./removeDuplicatePrimitives');
-var removeUnused = require('./removeUnused');
 var quantizeAttributes = require('./quantizeAttributes');
 var writeGltf = require('./writeGltf');
 var writeBinaryGltf = require('./writeBinaryGltf');
@@ -31,49 +30,38 @@ var writeSource = require('./writeSource');
 var defaultValue = Cesium.defaultValue;
 var defined = Cesium.defined;
 
-module.exports = {
-    processJSON : processJSON,
-    processJSONWithExtras : processJSONWithExtras,
-    processFile : processFile,
-    processJSONToDisk : processJSONToDisk,
-    processFileToDisk : processFileToDisk
-};
+var bakeAmbientOcclusion = AmbientOcclusion.bakeAmbientOcclusion;
+var removeAllUnused = RemoveUnusedElements.removeAllUnused;
 
-function writeSources(gltf) {
-    var embed = true;
-    var embedImage = true;
-    var writeSourcePromises = [
-        writeSource(gltf, undefined, 'buffers', embed, embedImage),
-        writeSource(gltf, undefined, 'images', embed, embedImage),
-        writeSource(gltf, undefined, 'shaders', embed, embedImage)
-    ];
-    return Promise.all(writeSourcePromises)
-        .then(function() {
-            return gltf;
-        });
-}
+module.exports = Pipeline;
+
+/**
+ * Main optimization pipeline.
+ * @constructor
+ */
+function Pipeline() {}
 
 /**
  * Add pipeline extras and load uris, then process the gltf.
+ * Options are passed to loadGltfUris and processJSONWithExtras.
  *
  * @param {Object} gltf A javascript object holding a glTF hierarchy.
  * @param {Object} options Options to apply to stages during optimization.
  * @returns {Promise} A promise that resolves to the processed gltf.
  *
- * options is passed to the functions called by this one
  * @see loadGltfUris
- * @see processJSONWithExtras
+ * @see Pipeline.processJSONWithExtras
  */
-function processJSON(gltf, options) {
+Pipeline.processJSON = function(gltf, options) {
     addPipelineExtras(gltf);
     return loadGltfUris(gltf, options)
         .then(function(gltf) {
-            return processJSONWithExtras(gltf, options);
+            return Pipeline.processJSONWithExtras(gltf, options);
         })
         .then(function(gltf) {
             return writeSources(gltf);
         });
-}
+};
 
 /**
  * Process a gltf that already has extras and loaded uris.
@@ -86,13 +74,9 @@ function processJSON(gltf, options) {
  * @param {Object} [options.quantize] Flag to run quantizeAttributes stage.
  * @returns {Promise} A promise that resolves to the processed gltf.
  */
-function processJSONWithExtras(gltfWithExtras, options) {
-    var stats = new OptimizationStatistics();
+Pipeline.processJSONWithExtras  = function(gltfWithExtras, options) {
     addDefaults(gltfWithExtras, options);
-    removeUnused(gltfWithExtras, stats);
-    if (options.printStats) {
-        printStats(stats);
-    }
+    removeAllUnused(gltfWithExtras);
     // It is generally better to merge the duplicate vertices before merging accessors.
     // Once accessors merge, there is more likely to be overlap in accessor usage between primitives
     // which limits the effectiveness of merging duplicate vertices.
@@ -108,7 +92,7 @@ function processJSONWithExtras(gltfWithExtras, options) {
     // Merging duplicate vertices again to prevent repeat data in newly combined primitives
     // mergeDuplicateVertices(gltfWithExtras);
     cacheOptimization(gltfWithExtras);
-    
+
     // run AO after cacheOptimization since AO adds new attributes.
     var aoOptions = options.aoOptions;
     if (defined(aoOptions)) {
@@ -116,8 +100,8 @@ function processJSONWithExtras(gltfWithExtras, options) {
         bakeAmbientOcclusion(aoOptions);
     }
 
-    // Run removeUnused stage again after all pipeline stages have been run to remove objects that become unused
-    removeUnused(gltfWithExtras);
+    // Run removeAllUnused stage again after all pipeline stages have been run to remove objects that become unused
+    removeAllUnused(gltfWithExtras);
     var waitForStages = [new Promise(function(resolve) {resolve();})];
     if (options.encodeNormals) {
         waitForStages.push(octEncodeNormals(gltfWithExtras));
@@ -146,16 +130,89 @@ function processJSONWithExtras(gltfWithExtras, options) {
             // Remove duplicates again after all stages to minimize the buffer size
             mergeDuplicateVertices(gltfWithExtras);
             return encodeImages(gltfWithExtras);
-    });
-}
+        });
+};
 
-function processFile(inputPath, options) {
+/**
+ * Process a gltf on disk into memory.
+ * Options are passed to processJSONWithExtras.
+ *
+ * @param {String} inputPath The input file path.
+ * @param {Object} options Options to apply tos tages during optimization.
+ * @returns {Object} The processed gltf.
+ *
+ * @see Pipeline.processJSONWithExtras
+ */
+Pipeline.processFile =function processFile(inputPath, options) {
     return readGltf(inputPath, options)
         .then(function(gltf) {
-            return processJSONWithExtras(gltf, options);
+            return Pipeline.processJSONWithExtras(gltf, options);
         })
         .then(function(gltf) {
             return writeSources(gltf);
+        });
+};
+
+/**
+ * Process a gltf in memory and writes it out to disk.
+ * Options are passed to loadGltfUris, processJSONWithExtras, writeGltf, and writeBinaryGltf.
+ *
+ * @param {Object} gltf A javascript object holding a glTF hierarchy.
+ * @param {String} outputPath The output file destination.
+ * @param {Object} options Options to apply to stages during optimization.
+ * @returns {Promise} A promise that resolves when the operation is complete.
+ *
+ * @see loadGltfUris
+ * @see Pipeline.processJSONWithExtras
+ * @see writeGltf
+ * @see writeBinaryGltf
+ */
+Pipeline.processJSONToDisk = function(gltf, outputPath, options) {
+    addPipelineExtras(gltf);
+    return loadGltfUris(gltf, options)
+        .then(function(gltf) {
+            return Pipeline.processJSONWithExtras(gltf, options);
+        })
+        .then(function(gltf) {
+            return writeFile(gltf, outputPath, options);
+        });
+};
+
+/**
+ * Processes a gltf on disk and writes it out to disk.
+ * Options are passed to processJSONWithExtras, readGltf, writeGltf, and writeBinaryGltf.
+ *
+ * @param {String} inputPath The input file path.
+ * @param {String} outputPath The output file destination
+ * @param {Object} options Options to apply to stages during optimization.
+ * @returns {Promise} A promise that resolves when the operation is complete.
+ *
+ * @see Pipeline.processJSONWithExtras
+ * @see readGltf
+ * @see writeGltf
+ * @see writeBinaryGltf
+ */
+Pipeline.processFileToDisk = function(inputPath, outputPath, options) {
+    return readGltf(inputPath, options)
+        .then(function(gltf) {
+            return Pipeline.processJSONWithExtras(gltf, options);
+        })
+        .then(function(gltf) {
+            return writeFile(gltf, outputPath, options);
+        });
+};
+
+function writeSources(gltf) {
+    var embed = true;
+    var embedImage = true;
+    var writeSourcePromises = [
+        writeSource(gltf, undefined, 'buffers', embed, embedImage),
+        writeSource(gltf, undefined, 'images', embed, embedImage),
+        writeSource(gltf, undefined, 'shaders', embed, embedImage)
+    ];
+    return Promise.all(writeSourcePromises)
+        .then(function() {
+            return gltf;
         });
 }
 
@@ -176,55 +233,6 @@ function writeFile(gltf, outputPath, options) {
         return writeBinaryGltf(gltf, writeOptions);
     }
     return writeGltf(gltf, writeOptions);
-}
-
-/**
- * Process a gltf in memory and writes it out to disk.
- *
- * @param {Object} gltf A javascript object holding a glTF hierarchy.
- * @param {String} outputPath The output file destination.
- * @param {Object} options Options to apply to stages during optimization.
- * @returns {Promise} A promise that resolves when the operation is complete.
- *
- * options is passed to the functions called by this one
- * @see loadGltfUris
- * @see processJSONWithExtras
- * @see writeGltf
- * @see writeBinaryGltf
- */
-function processJSONToDisk(gltf, outputPath, options) {
-    addPipelineExtras(gltf);
-    return loadGltfUris(gltf, options)
-        .then(function(gltf) {
-            return processJSONWithExtras(gltf, options);
-        })
-        .then(function(gltf) {
-            return writeFile(gltf, outputPath, options);
-        });
-}
-
-/**
- * Processes a gltf on disk and writes it out to disk.
- *
- * @param {String} inputPath The input file path.
- * @param {String} outputPath The output file destination
- * @param {Object} options Options to apply to stages during optimization.
- * @returns {Promise} A promise that resolves when the operation is complete.
- *
- * options is passed to the functions called by this one
- * @see processJSONWithExtras
- * @see readGltf
- * @see writeGltf
- * @see writeBinaryGltf
- */
-function processFileToDisk(inputPath, outputPath, options) {
-    return readGltf(inputPath, options)
-        .then(function(gltf) {
-            return processJSONWithExtras(gltf, options);
-        })
-        .then(function(gltf) {
-           return writeFile(gltf, outputPath, options);
-        });
 }
 
 function printStats(stats) {
