+++ lib/StaticUniformGrid.js
@@ -16,6 +16,7 @@ var halfDimensionsScratch = new Cartesian3();
 /**
  * A Uniform Grid for unmoving items.
  * @alias StaticUniformGrid
+ * @private
  * @constructor
  *
  * @param {Object[]} items Items that the grid should store.
