+++ lib/combineNodes.js
@@ -11,7 +11,7 @@ var AccessorReader = require('./AccessorReader');
 var getUniqueId = require('./getUniqueId');
 var NodeHelpers = require('./NodeHelpers');
 var readAccessor = require('./readAccessor');
-var removeUnusedNodes = require('./removeUnusedNodes');
+var removeUnusedNodes = require('./RemoveUnusedElements').removeUnusedNodes;
 var writeAccessor = require('./writeAccessor');
 
 module.exports = combineNodes;
