+++ lib/loadGltfUris.js
@@ -19,13 +19,13 @@ module.exports = loadGltfUris;
 /**
  * Load uris in the glTF into buffers.
  * The buffer data is placed into extras._pipeline.source for the corresponding object.
+ * glTF must be initialized for the pipeline.
  *
  * @param {Object} gltf A javascript object holding a glTF hierarchy.
  * @param {Object} options Options for loading uris.
  * @param {String} options.basePath The path to look in when loading external files.
  * @returns {Promise} A promise that resolves to gltf when all uris are loaded.
  *
- * gltf must be initialized for the pipeline.
  * @see addPipelineExtras
  */
 function loadGltfUris(gltf, options) {
@@ -91,7 +91,9 @@ function loadURI(gltf, basePath, name) {
 }
 
 /**
- * Return the extension of the data uri
+ * Return the extension of the data uri.
+ *
+ * @private
  * @param {String} uri
  * @returns {String}
  */
@@ -130,8 +132,9 @@ function readImageFromFile(object, name, uriPath) {
 }
 
 /**
- * Generate a jimp image for png, jpeg, or bmp
+ * Generate a jimp image for png, jpeg, or bmp.
  *
+ * @private
  * @param {Object} object
  * @returns {Promise}
  */
