+++ lib/removeObject.js
@@ -7,13 +7,12 @@ module.exports = removeObject;
 /**
  * @private
  */
-function removeObject(gltf, name, usedId, stats) {
+function removeObject(gltf, name, usedId) {
     // Iterate through objects and remove those that are not in the hash
     var numberRemoved = 0;
     var objects = gltf[name];
     if (defined(objects)) {
         var used = {};
-
         for (var id in objects) {
             if (objects.hasOwnProperty(id)) {
                 // If this object is in the hash, then keep it in the glTF asset
@@ -24,13 +23,7 @@ function removeObject(gltf, name, usedId, stats) {
                 }
             }
         }
-
-        if (defined(stats)) {
-            stats.numberRemoved[name] += numberRemoved;
-        }
-
         gltf[name] = used;
     }
-
     return gltf;
 }
