+++ lib/triangleAxisAlignedBoundingBoxOverlap.js
@@ -1,10 +1,5 @@
 'use strict';
 
-/**
- * Function and helpers for computing the intersection between a static triangle and a static AABB.
- * Adapted from Tomas Akenine-Möller's public domain implementation: http://www.cs.lth.se/home/Tomas_Akenine_Moller/code/
- */
-
 var Cesium = require('cesium');
 var Cartesian3 = Cesium.Cartesian3;
 
@@ -21,6 +16,11 @@ var v2 = new Cartesian3();
 var triangleNormalScratch = new Cartesian3();
 var halfHeightScratch = new Cartesian3();
 
+/**
+ * Function and helpers for computing the intersection between a static triangle and a static AABB.
+ * Adapted from Tomas Akenine-Möller's public domain implementation: http://www.cs.lth.se/home/Tomas_Akenine_Moller/code/
+ * @private
+ */
 function triangleAxisAlignedBoundingBoxOverlap(axisAlignedBoundingBox, triangle) {
     // Use the separating axis theorem to test overlap between triangle and box.
     // Need to test for overlap in these directions:
