+++ specs/lib/AmbientOcclusionSpec.js
@@ -6,7 +6,7 @@ var CesiumMath = Cesium.Math;
 var Cartesian3 = Cesium.Cartesian3;
 var Matrix4 = Cesium.Matrix4;
 
-var bakeAmbientOcclusion = require('../../lib/bakeAmbientOcclusion');
+var AmbientOcclusion = require('../../lib/AmbientOcclusion');
 var NodeHelpers = require('../../lib/NodeHelpers');
 var readGltf = require('../../lib/readGltf');
 var StaticUniformGrid = require('../../lib/StaticUniformGrid');
@@ -30,7 +30,7 @@ function cloneGltfWithJimps(gltf) {
     return gltfClone;
 }
 
-describe('bakeAmbientOcclusion', function() {
+describe('AmbientOcclusion', function() {
     var boxOverGroundGltf;
 
     var indices = [0,1,2,0,2,3];
@@ -235,7 +235,7 @@ describe('bakeAmbientOcclusion', function() {
             gltfWithExtras : testGltf,
             rayDistance : 10.0
         };
-        var raytracerScene = bakeAmbientOcclusion.generateRaytracerScene(options);
+        var raytracerScene = AmbientOcclusion.generateRaytracerScene(options);
         var triangleSoup = raytracerScene.triangleGrid.items;
 
         // because of the uniform scale, expect triangles to be bigger
@@ -276,7 +276,7 @@ describe('bakeAmbientOcclusion', function() {
             sceneID : testGltf.scene,
             gltfWithExtras : testGltf
         };
-        var raytracerScene = bakeAmbientOcclusion.generateRaytracerScene(options);
+        var raytracerScene = AmbientOcclusion.generateRaytracerScene(options);
         var triangleSoup = raytracerScene.triangleGrid.items;
 
         // ground plane size is based on the near culling distance, scene size, and maximum ray depth.
@@ -323,7 +323,7 @@ describe('bakeAmbientOcclusion', function() {
 
         for (i = 0; i < 6; i++) {
             var texel = texelPoints[i];
-            samples[i] = bakeAmbientOcclusion.computeAmbientOcclusionAt({
+            samples[i] = AmbientOcclusion.computeAmbientOcclusionAt({
                 position : texel.position,
                 normal : texel.normal,
                 numberRays : 16,
@@ -362,7 +362,7 @@ describe('bakeAmbientOcclusion', function() {
 
         for (var i = 0; i < 3; i++) {
             var texel = texelPoints[i];
-            samples[i] += bakeAmbientOcclusion.computeAmbientOcclusionAt({
+            samples[i] += AmbientOcclusion.computeAmbientOcclusionAt({
                 position : texel.position,
                 normal : texel.normal,
                 numberRays : 16,
@@ -386,7 +386,7 @@ describe('bakeAmbientOcclusion', function() {
             toTexture: true,
             gltfWithExtras: boxOverGroundGltfClone
         };
-        bakeAmbientOcclusion.bakeAmbientOcclusion(options);
+        AmbientOcclusion.bakeAmbientOcclusion(options);
 
         expect(Object.keys(boxOverGroundGltfClone.images).length).toEqual(2);
         expect(Object.keys(boxOverGroundGltfClone.textures).length).toEqual(2);
@@ -415,7 +415,7 @@ describe('bakeAmbientOcclusion', function() {
             toTexture: true,
             gltfWithExtras: boxOverGroundGltfClone
         };
-        bakeAmbientOcclusion.bakeAmbientOcclusion(options);
+        AmbientOcclusion.bakeAmbientOcclusion(options);
 
         expect(Object.keys(boxOverGroundGltfClone.images).length).toEqual(2);
         expect(Object.keys(boxOverGroundGltfClone.textures).length).toEqual(2);
@@ -444,7 +444,7 @@ describe('bakeAmbientOcclusion', function() {
             toTexture: true,
             gltfWithExtras: boxOverGroundGltfClone
         };
-        bakeAmbientOcclusion.bakeAmbientOcclusion(options);
+        AmbientOcclusion.bakeAmbientOcclusion(options);
 
         expect(Object.keys(boxOverGroundGltfClone.images).length).toEqual(3); // 1 unused image and 2 images with AO
         expect(Object.keys(boxOverGroundGltfClone.textures).length).toEqual(2);
@@ -473,7 +473,7 @@ describe('bakeAmbientOcclusion', function() {
             toTexture: true,
             gltfWithExtras: boxOverGroundGltfClone
         };
-        bakeAmbientOcclusion.bakeAmbientOcclusion(options);
+        AmbientOcclusion.bakeAmbientOcclusion(options);
 
         expect(Object.keys(boxOverGroundGltfClone.images).length).toEqual(3); // 1 unused image and 2 with AO
         expect(Object.keys(boxOverGroundGltfClone.textures).length).toEqual(3); // 1 unused texture, 2 with AO
@@ -500,7 +500,7 @@ describe('bakeAmbientOcclusion', function() {
             toTexture: true,
             gltfWithExtras: boxOverGroundGltfClone
         };
-        bakeAmbientOcclusion.bakeAmbientOcclusion(options);
+        AmbientOcclusion.bakeAmbientOcclusion(options);
 
         expect(Object.keys(boxOverGroundGltfClone.images).length).toEqual(3); // 1 unused image and 2 images with AO
         expect(Object.keys(boxOverGroundGltfClone.textures).length).toEqual(3); // 1 unused texture, 2 with AO
@@ -515,7 +515,7 @@ describe('bakeAmbientOcclusion', function() {
             toVertex: true,
             gltfWithExtras: boxOverGroundGltfClone
         };
-        bakeAmbientOcclusion.bakeAmbientOcclusion(options);
+        AmbientOcclusion.bakeAmbientOcclusion(options);
 
         expect(Object.keys(boxOverGroundGltfClone.accessors).length).toEqual(10);
         var cubeMeshPrimitives = boxOverGroundGltfClone.meshes.Cube_mesh.primitives;
@@ -534,7 +534,7 @@ describe('bakeAmbientOcclusion', function() {
             toVertex: true,
             gltfWithExtras: boxOverGroundGltfClone
         };
-        bakeAmbientOcclusion.bakeAmbientOcclusion(options);
+        AmbientOcclusion.bakeAmbientOcclusion(options);
 
         expect(Object.keys(boxOverGroundGltfClone.materials).length).toEqual(4);
         expect(Object.keys(boxOverGroundGltfClone.techniques).length).toEqual(2);
@@ -586,7 +586,7 @@ describe('bakeAmbientOcclusion', function() {
             }
         };
 
-        bakeAmbientOcclusion.raytraceAtTriangleCenters(primitive, 'meshPrimitiveID', parameters, node);
+        AmbientOcclusion.raytraceAtTriangleCenters(primitive, 'meshPrimitiveID', parameters, node);
 
         var samples = aoBuffer.samples;
         var counts = aoBuffer.count;
@@ -636,7 +636,7 @@ describe('bakeAmbientOcclusion', function() {
             }
         };
 
-        bakeAmbientOcclusion.raytraceOverTriangleSamples(primitive, 'meshPrimitiveID', parameters, node);
+        AmbientOcclusion.raytraceOverTriangleSamples(primitive, 'meshPrimitiveID', parameters, node);
 
         var samples = aoBuffer.samples;
         var counts = aoBuffer.count;
@@ -655,7 +655,7 @@ describe('bakeAmbientOcclusion', function() {
             gltfWithExtras: boxOverGroundGltf
         };
 
-        var options = bakeAmbientOcclusion.generateOptions(aoOptions);
+        var options = AmbientOcclusion.generateOptions(aoOptions);
         expect(options.toTexture).toEqual(false);
         expect(options.groundPlane).toEqual(false);
         expect(options.ambientShadowContribution).toEqual(0.5);
@@ -672,7 +672,7 @@ describe('bakeAmbientOcclusion', function() {
             quality: 'medium',
             gltfWithExtras: boxOverGroundGltf
         };
-        options = bakeAmbientOcclusion.generateOptions(aoOptions);
+        options = AmbientOcclusion.generateOptions(aoOptions);
         expect(options.toTexture).toEqual(false);
         expect(options.groundPlane).toEqual(false);
         expect(options.ambientShadowContribution).toEqual(0.5);
@@ -689,7 +689,7 @@ describe('bakeAmbientOcclusion', function() {
             quality: 'high',
             gltfWithExtras: boxOverGroundGltf
         };
-        options = bakeAmbientOcclusion.generateOptions(aoOptions);
+        options = AmbientOcclusion.generateOptions(aoOptions);
         expect(options.toTexture).toEqual(false);
         expect(options.groundPlane).toEqual(false);
         expect(options.ambientShadowContribution).toEqual(0.5);
@@ -711,7 +711,7 @@ describe('bakeAmbientOcclusion', function() {
             rayDistance: 10.0,
             gltfWithExtras: boxOverGroundGltf
         };
-        var options = bakeAmbientOcclusion.generateOptions(aoOptions);
+        var options = AmbientOcclusion.generateOptions(aoOptions);
         expect(options.toTexture).toEqual(true);
         expect(options.groundPlane).toEqual(false);
         expect(options.ambientShadowContribution).toEqual(0.5);
@@ -745,7 +745,7 @@ describe('bakeAmbientOcclusion', function() {
             functionArguments : ['arg0', 'arg1', 'arg2']
         };
 
-        var extractedCommand = bakeAmbientOcclusion.extractInstructionWithFunctionCall(options);
+        var extractedCommand = AmbientOcclusion.extractInstructionWithFunctionCall(options);
         expect(extractedCommand).toEqual('val = command(arg0, arg1 * (arg2 + innerCommand(innerArg0))) + 0.0');
     });
 
@@ -769,7 +769,7 @@ describe('bakeAmbientOcclusion', function() {
             snippet : 'innerCommand'
         };
 
-        bakeAmbientOcclusion.injectGlslAfterInstructionContaining(options);
+        AmbientOcclusion.injectGlslAfterInstructionContaining(options);
         var newSource = shader.extras._pipeline.source.toString();
         expect(newSource).toEqual('function(arg0, arg1, arg2, arg3, innerArg0) {' +
             'command(arg3);' +
